Fixed tests broken from restructuring
