+++ django_swagger_tester/__init__.py
@@ -1,2 +1,2 @@
-__version__ = '0.2.3'
+__version__ = '0.3.0'
 __author__ = 'Sondre Lillebø Gundersen'
