+++ tests/test_response_validation/functional/test_generated_static_schema.py
