+++ tests/test_response_validation/unit/static_schema/test_load_schema_file.py
@@ -4,7 +4,7 @@ import pytest
 from django.conf import settings, settings as django_settings
 from django.core.exceptions import ImproperlyConfigured
 
-from django_swagger_tester.static_schema.base import LoadStaticSchema
+from django_swagger_tester.static_schema.loader import LoadStaticSchema
 
 yml_path = django_settings.BASE_DIR + '/demo_project/openapi-schema.yml'
 json_path = django_settings.BASE_DIR + '/demo_project/openapi-schema.json'
@@ -16,7 +16,7 @@ def test_successful_yml_fetch(monkeypatch) -> None:
     """
     monkeypatch.setattr(django_settings, 'SWAGGER_TESTER', {'PATH': yml_path})
     base = LoadStaticSchema('api/v1/trucks/correct', 'GET', status_code=200)
-    content = base.load_schema_file()
+    content = base.get_schema()
     assert 'openapi' in content
 
 
@@ -26,7 +26,7 @@ def test_successful_json_fetch(monkeypatch) -> None:
     """
     monkeypatch.setattr(django_settings, 'SWAGGER_TESTER', {'PATH': json_path})
     base = LoadStaticSchema('api/v1/trucks/correct', 'GET', status_code=200)
-    content = base.load_schema_file()
+    content = base.get_schema()
     assert 'title' in content
 
 
@@ -40,7 +40,7 @@ def test_non_existent_file(caplog, monkeypatch) -> None:
         ImproperlyConfigured,
         match='The path `test` does not point to a valid file. Make sure to point to the specification file.',
     ):
-        base.load_schema_file()
+        base.get_schema()
         assert 'Path `test` does not resolve as a valid file.' in caplog.records
 
 
@@ -52,14 +52,14 @@ def test_unreadable_file(monkeypatch, caplog) -> None:
     def mocked_isfile(*args, **kwargs):
         return True
 
-    monkeypatch.setattr('django_swagger_tester.static_schema.load_schema.os.path.isfile', mocked_isfile)
+    monkeypatch.setattr('django_swagger_tester.static_schema.loader.os.path.isfile', mocked_isfile)
 
     monkeypatch.setattr(django_settings, 'SWAGGER_TESTER', {'PATH': yml_path + 's'})
     base = LoadStaticSchema('api/v1/trucks/correct', 'GET', status_code=200)
     with pytest.raises(
         ImproperlyConfigured, match='Unable to read the schema file. Please make sure the path setting is correct.'
     ):
-        base.load_schema_file()
+        base.get_schema()
 
 
 def test_bad_filetype(monkeypatch) -> None:
@@ -71,4 +71,4 @@ def test_bad_filetype(monkeypatch) -> None:
     with pytest.raises(
         ImproperlyConfigured, match='The specified file path does not seem to point to a JSON or YAML file.'
     ):
-        base.load_schema_file()
+        base.get_schema()
