+++ tests/test_version.py
@@ -2,4 +2,4 @@ from django_swagger_tester import __version__
 
 
 def test_version():
-    assert __version__ == '0.2.3'
+    assert __version__ == '0.3.0'
